♻️ BirthdayyClient class & abstracted options
